+++ tkpdseller/src/main/java/com/tokopedia/seller/goldmerchant/statistic/domain/mapper/GMTransactionStatDomainMapper.java
@@ -189,7 +189,7 @@ public class GMTransactionStatDomainMapper implements Func1<GetTransactionGraph,
         try {
             gmTopAdsAmountViewModel.percentage = gmTopAdsAmountViewModel.amount - previousAmount / gmTopAdsAmountViewModel.amount;
         } catch (Exception e) {
-            Log.e(TAG, String.format("amount %s previous amount %s", Integer.toString(gmTopAdsAmountViewModel.amount), Long.toString(previousAmount)));
+            Log.e(TAG, String.format("amount %s previous amount %s", Long.toString(gmTopAdsAmountViewModel.amount), Long.toString(previousAmount)));
             gmTopAdsAmountViewModel.percentage = -GMStatConstant.NO_DATA_AVAILABLE;
         }
